+++ app/src/components/HelpView.tsx
@@ -1,4 +1,6 @@
+import { mergeClassNames } from "@/utils/className";
 import { TW_COLORS } from "@/utils/constants";
+import { typography } from "@/utils/typography";
 import { View, Text } from "react-native";
 
 interface Props {
@@ -16,6 +18,8 @@ const HelpView = ({ title, description, link }: Props) => {
       <Text className="text-base mb-4 leading-6" style={{ color: TW_COLORS.TEXT_SECONDARY }}>
         {description}
       </Text>
+      <Text className={mergeClassNames(typography.textXlBold, "mb-4 text-cnam-primary-950")}>{title}</Text>
+      <Text className={mergeClassNames(typography.textLgRegular, "text-cnam-primary-950")}>{description}</Text>
       {link && (
         <Text className="text-base text-center mb-4 leading-6" style={{ color: TW_COLORS.TEXT_SECONDARY }}>
           {link}