+++ app/src/components/SelectionnableItem.tsx
@@ -13,8 +13,10 @@ type DifficultyOptionProps = {
   selected: boolean;
   onPress: (id: string | number) => void;
   className?: string;
+  disabled?: boolean;
   icon?: React.ComponentType<any>;
   boxPosition?: string;
+  shape?: "square" | "circle";
 };
 
 export default function SelectionnableItem({ id, label, description, selected, onPress, className, icon, boxPosition }: DifficultyOptionProps) {
@@ -58,16 +60,29 @@ export default function SelectionnableItem({ id, label, description, selected, o
   );
 }
 
-export function LightSelectionnableItem({ id, label, description, selected, onPress, className, icon, boxPosition }: DifficultyOptionProps) {
+export function LightSelectionnableItem({ id, label, description, selected, onPress, className, disabled, icon, shape }: DifficultyOptionProps) {
   return (
-    <TouchableOpacity onPress={() => onPress(id)} className={mergeClassNames("mb-3 py-3 border-b", "border-gray-300 bg-transparent", className)}>
+    <TouchableOpacity
+      disabled={disabled}
+      onPress={() => onPress(id)}
+      className={mergeClassNames("mb-3 py-3 border-b", "border-gray-300 bg-transparent", className, disabled ? "opacity-50" : "")}
+    >
       <View className={mergeClassNames("flex-row items-center", boxPosition === "top" ? "items-start" : "items-center")}>
         {selected ? (
-          <View className="w-6 h-6 rounded-md items-center justify-center bg-cnam-primary-800 mr-4">
-            <Text className="text-white text-base font-bold">✓</Text>
+          <View
+            className={mergeClassNames("w-6 h-6 items-center justify-center bg-primary mr-4", shape === "circle" ? "rounded-full" : "rounded-md")}
+          >
+            { shape == "circle" ?
+            <View className="bg-white w-2 h-2 rounded-xl"/>
+            :<Text className="text-white text-base font-bold">✓</Text>}
           </View>
         ) : (
-          <View className="w-6 h-6 rounded-md items-center justify-center border-2 border-gray-300 mr-4">
+          <View
+            className={mergeClassNames(
+              "w-6 h-6 items-center justify-center border-2 border-gray-300 mr-4",
+              shape === "circle" ? "rounded-full" : "rounded-md"
+            )}
+          >
             <Text className="text-white text-xs" />
           </View>
         )}
@@ -77,7 +92,7 @@ export function LightSelectionnableItem({ id, label, description, selected, onPr
           </View>
         )}
         <View className="flex-1">
-          <Text className={mergeClassNames(typography.textMdMedium, "text-cnam-primary-900")}>{label}</Text>
+          <Text className={mergeClassNames(typography.textMdMedium, "text-brand-950", disabled ? "line-through" : "")}>{label}</Text>
           {description && <Text className={mergeClassNames(typography.textSmMedium, "text-gray-600 mt-1")}>{description}</Text>}
         </View>
       </View>
@@ -85,8 +100,29 @@ export function LightSelectionnableItem({ id, label, description, selected, onPr
   );
 }
 
-export function InputSelectionnableItem({ id, label, selected, onPress, className, icon }: DifficultyOptionProps) {
+export function InputSelectionnableItem({
+  id,
+  label,
+  selected,
+  onPress,
+  className,
+  icon,
+  shape,
+  validationError,
+  onTextChange,
+}: DifficultyOptionProps & {
+  validationError?: string;
+  onTextChange?: (text: string) => void;
+}) {
   const [value, setValue] = useState<string>("");
+
+  const handleTextChange = (text: string) => {
+    setValue(text);
+    if (onTextChange) {
+      onTextChange(text);
+    }
+  };
+
   return (
     <TouchableOpacity
       key={id}
@@ -95,11 +131,20 @@ export function InputSelectionnableItem({ id, label, selected, onPress, classNam
     >
       <View className="flex-row items-center">
         {selected ? (
-          <View className="w-6 h-6 rounded-md items-center justify-center bg-cnam-primary-800 mr-2">
-            <Text className="text-white text-base font-bold">✓</Text>
+          <View
+            className={mergeClassNames("w-6 h-6 items-center justify-center bg-primary mr-4", shape === "circle" ? "rounded-full" : "rounded-md")}
+          >
+            { shape == "circle" ?
+            <View className="bg-white w-2 h-2 rounded-xl"/>
+            :<Text className="text-white text-base font-bold">✓</Text>}
           </View>
         ) : (
-          <View className="w-6 h-6 rounded-md items-center justify-center border-2 border-gray-300 mr-4">
+          <View
+            className={mergeClassNames(
+              "w-6 h-6 items-center justify-center border-2 border-gray-300 mr-4",
+              shape === "circle" ? "rounded-full" : "rounded-md"
+            )}
+          >
             <Text className="text-white text-xs" />
           </View>
         )}
@@ -112,7 +157,7 @@ export function InputSelectionnableItem({ id, label, selected, onPress, classNam
           <Text className={mergeClassNames(typography.textSmMedium, "text-gray-800 mb-2")}>{label}</Text>
           <View className="rounded rounded-lg flex-1">
             <TextInput
-              onChangeText={(text) => setValue(text)}
+              onChangeText={handleTextChange}
               className={mergeClassNames(typography.textMdRegular, "text-left border border-gray-300 p-2 rounded rounded-lg flex-1")}
             />
             <TouchableOpacity
@@ -126,6 +171,7 @@ export function InputSelectionnableItem({ id, label, selected, onPress, classNam
               <Text className={mergeClassNames(typography.textMdSemibold, "text-brand-800")}>Valider</Text>
             </TouchableOpacity>
           </View>
+          {validationError && <Text className={mergeClassNames(typography.textSmMedium, "text-red-600 mt-2")}>{validationError}</Text>}
         </View>
       </View>
     </TouchableOpacity>